Prevent file, swift+config and filesystem schemes
This change ensures that 'file', 'filesystem', and 'swift+config' URI
schemes are not allowed when setting the location field. A previous
fix to CVE-2014-9493 attempted to address this issue but did not
include 'filesystem', a URI scheme allowed by the glance_store.

Without this fix in place it is possible for a client to access any file
the glance-api server has read permissions for.

(cherry picked from commit 5191ed1879c5fd5b2694f922bcedec232f461088)

Conflicts:
	glance/common/store_utils.py

Change-Id: I02cd099a8634b9c7e3cf8f172bcbd33f8edcbc83
Closes-Bug: #1408663

Conflicts:
	glance/store/__init__.py
	glance/tests/unit/test_store_location.py
	glance/tests/unit/v1/test_api.py